+++ src/tests.rs
@@ -164,4 +164,67 @@ mod tests {
             }
         }
     }
+
+    #[cfg(feature = "ipp")]
+    mod ipp_comparison_tests {
+        use crate::{fft_convolver::ipp_fft, fft_convolver::rust_fft, Convolution};
+
+        fn compare_implementations(impulse_response: &[f32], input: &[f32], block_size: usize) {
+            let max_len = impulse_response.len();
+
+            let mut rust_convolver =
+                rust_fft::FFTConvolver::init(impulse_response, block_size, max_len);
+            let mut ipp_convolver =
+                ipp_fft::FFTConvolver::init(impulse_response, block_size, max_len);
+
+            let mut rust_output = vec![0.0; input.len()];
+            let mut ipp_output = vec![0.0; input.len()];
+
+            rust_convolver.process(input, &mut rust_output);
+            ipp_convolver.process(input, &mut ipp_output);
+
+            for i in 0..input.len() {
+                assert!(
+                    (rust_output[i] - ipp_output[i]).abs() < 1e-5,
+                    "Outputs differ at position {}: rust={}, ipp={}",
+                    i,
+                    rust_output[i],
+                    ipp_output[i]
+                );
+            }
+        }
+
+        #[test]
+        fn test_ipp_vs_rust_impulse() {
+            let mut response = vec![0.0; 1024];
+            response[0] = 1.0;
+            let input = vec![1.0; 1024];
+
+            compare_implementations(&response, &input, 256);
+        }
+
+        #[test]
+        fn test_ipp_vs_rust_decay() {
+            let mut response = vec![0.0; 1024];
+            for i in 0..response.len() {
+                response[i] = 0.9f32.powi(i as i32);
+            }
+            let input = vec![1.0; 1024];
+
+            compare_implementations(&response, &input, 256);
+        }
+
+        #[test]
+        fn test_ipp_vs_rust_sine() {
+            let mut response = vec![0.0; 1024];
+            response[0] = 1.0;
+
+            let mut input = vec![0.0; 1024];
+            for i in 0..input.len() {
+                input[i] = (i as f32 * 0.1).sin();
+            }
+
+            compare_implementations(&response, &input, 128);
+        }
+    }
 }
